some fix to make it more c++

+++ src/learner/objective-inl.hpp
@@ -6,9 +6,9 @@
  * \author Tianqi Chen, Kailong Chen
  */
 #include <vector>
-#include <cmath>
 #include <algorithm>
 #include <utility>
+#include <cmath>
 #include <functional>
 #include "../data.h"
 #include "./objective.h"
@@ -37,7 +37,7 @@ struct LossType {
       case kLogisticRaw:
       case kLinearSquare: return x;
       case kLogisticClassify:
-      case kLogisticNeglik: return 1.0f / (1.0f + expf(-x));
+      case kLogisticNeglik: return 1.0f / (1.0f + std::exp(-x));
       default: utils::Error("unknown loss_type"); return 0.0f;
     }
   }
@@ -50,7 +50,7 @@ struct LossType {
   inline float FirstOrderGradient(float predt, float label) const {
     switch (loss_type) {
       case kLinearSquare: return predt - label;
-      case kLogisticRaw: predt = 1.0f / (1.0f + expf(-predt));
+      case kLogisticRaw: predt = 1.0f / (1.0f + std::exp(-predt));
       case kLogisticClassify:
       case kLogisticNeglik: return predt - label;
       default: utils::Error("unknown loss_type"); return 0.0f;
@@ -65,7 +65,7 @@ struct LossType {
   inline float SecondOrderGradient(float predt, float label) const {
     switch (loss_type) {
       case kLinearSquare: return 1.0f;
-      case kLogisticRaw: predt = 1.0f / (1.0f + expf(-predt));
+      case kLogisticRaw: predt = 1.0f / (1.0f + std::exp(-predt));
       case kLogisticClassify:
       case kLogisticNeglik: return predt * (1 - predt);
       default: utils::Error("unknown loss_type"); return 0.0f;
@@ -80,7 +80,7 @@ struct LossType {
         loss_type == kLogisticNeglik ) {
       utils::Check(base_score > 0.0f && base_score < 1.0f,
                    "base_score must be in (0,1) for logistic loss");
-      base_score = -logf(1.0f / base_score - 1.0f);
+      base_score = -std::log(1.0f / base_score - 1.0f);
     }
     return base_score;
   }
@@ -419,8 +419,8 @@ class LambdaRankObjNDCG : public LambdaRankObj {
       for (size_t i = 0; i < pairs.size(); ++i) {
         unsigned pos_idx = pairs[i].pos_index;
         unsigned neg_idx = pairs[i].neg_index;
-        float pos_loginv = 1.0f / logf(pos_idx + 2.0f);
-        float neg_loginv = 1.0f / logf(neg_idx + 2.0f);
+        float pos_loginv = 1.0f / std::log(pos_idx + 2.0f);
+        float neg_loginv = 1.0f / std::log(neg_idx + 2.0f);
         int pos_label = static_cast<int>(sorted_list[pos_idx].label);
         int neg_label = static_cast<int>(sorted_list[neg_idx].label);
         float original =
@@ -438,7 +438,7 @@ class LambdaRankObjNDCG : public LambdaRankObj {
     for (size_t i = 0; i < labels.size(); ++i) {
       const unsigned rel = static_cast<unsigned>(labels[i]);
       if (rel != 0) {
-        sumdcg += ((1 << rel) - 1) / logf(static_cast<float>(i + 2));
+        sumdcg += ((1 << rel) - 1) / std::log(static_cast<float>(i + 2));
       }
     }
     return static_cast<float>(sumdcg);
